Android: Code signing module
Android.sdk was already signing the apk package but it was using
hardcoded debug key and could only be used to run the application
locally.
The signing part of the android apk/aab package generation is now
in the new codesign module. By default the same debug key is used.
But it is now also possible to configure a different key.

Task-number: QBS-899
Change-Id: I49c54a4d55578c48363805e927392b3a468805f0